* Adding synchronization to all methods in the JCL-Store, so reading and
    writing operations are consistent in a multi-thread environment.

* Initial implementation of the ChainBranchInfo implementations, which
    will allow as to check the branch a Block belongs to in almost O(1)
    time

* Implementation of the getChainTips(blockHash) and
    getFirstBlockInPath(blockHash) methods

* Implementation of the "getOrphanBlocks()" method, and refactoring to
    move the "_automaticPrunning" methods t the interfaces so the
    LevelDB and FoundationDB are kept clean.

* adding calculation of the accumulate chainWork

* Implementation of Automatic Orphan Prunning and doc update

* camelCase fixed

* fixing a problem in the BlockChainInfo Serializer

* doc update

+++ store-keyValue/src/main/java/com/nchain/jcl/store/keyValue/blockChainStore/BlockChainInfoSerializer.java
@@ -33,7 +33,7 @@ public class BlockChainInfoSerializer {
         if (object == null) return null;
         ByteArrayWriter writer = new ByteArrayWriter();
         BitcoinSerializerUtils.serializeVarStr(object.getBlockHash(), writer);
-        BitcoinSerializerUtils.serializeVarInt(object.getChainWork().intValue(), writer);
+        BitcoinSerializerUtils.serializeVarInt(object.getChainWork().longValue(), writer);
         BitcoinSerializerUtils.serializeVarInt(object.getHeight(), writer);
         BitcoinSerializerUtils.serializeVarInt(object.getTotalChainSize(), writer);
         writer.writeUint32LE(object.getChainPathId());
